Removes redundant tests
