+++ project_admin/tests/tests_views.py
@@ -3,25 +3,53 @@ from django.core.management import call_command
 from django.conf import settings
 
 
-class User_Views_Test(TestCase):
+class AdminLoginTestCase(TestCase):
+    """
+    Tests Admin login
+    """
 
     def setUp(self):
         """
         Set up the app for following tests.
         """
         settings.DEBUG = True
-        settings.ADMIN_PASSWORD = '123'
+        settings.ADMIN_PASSWORD = 'test1234'
         call_command('init_proj_config')
 
-    def test_project_config_view(self):
-        response = self.client.get("/project-admin/")
+    def test_admin_login_redirect(self):
+        """
+        Test redirection to the login page
+        """
+        c = Client()
+        response = c.get("/project-admin/")
         self.assertRedirects(response, '/project-admin/login')
-        self.assertEqual(response.status_code, 302)
+
+    def test_admin_login_success(self):
+        """
+        Test successful log in attempt.
+        """
+        c = Client()
+        response = c.post("/project-admin/login/",
+                          {'password': 'test1234'},
+                          follow=True)
+        self.assertRedirects(response, '/project-admin/')
+        self.assertContains(response, "Current Project Configuration")
+
+    def test_admin_login_fail(self):
+        """
+        Test unsuccessful login attempt.
+        """
+        c = Client()
+        response = c.post("/project-admin/login/", {'password': 'meep'})
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, "Password incorrect")
 
     def test_project_config_page_returns_correct_html(self):
         c = Client()
-        response = c.post("/project-admin/login", 
-                   {'user': 'admin', 'password': '123'}, follow=True)
-        self.assertContains(response, '<h1>Current Project Configuration</h1>', 
-                   status_code=200)
+        response = c.post("/project-admin/login",
+                          {'password': 'test1234'},
+                          follow=True)
+        self.assertContains(response,
+                            '<h1>Current Project Configuration</h1>',
+                            status_code=200)
         self.assertTemplateUsed(response, 'project_admin/home.html')
